+++ Changelog.md
@@ -2,6 +2,17 @@
 
 ## Upcoming Changes
 
+## Version 1.15
+
+Add flag `--force-empty-dir-permissions <3-octal-values>` that force sets the
+permissions for all extracted empty directories.
+
+Changes behavior to force set empty dir permissions on archive creation only if
+`--force-empty-dir-permissions <...>` is used instead of
+`--force-dir-permissions <...>`.
+
+Internal refactorings/fixes.
+
 ## Version 1.14
 
 Fix not setting empty directory permissions when extracted.